document_processor/requirements/dev.txt
-r base.txt

django-debug-toolbar
django-silk
markdownify
openai
pdfminer
pdfminer.six
pdfplumber
pika
PyMuPDF 
pypdf
ipython
pylint
pytest
pytest-django
black
flake8
ipdb
isort
coverage
model-bakery